+++ generic/query_generic.go
@@ -6,51 +6,48 @@ import (
 	"github.com/mlange-42/arche/ecs"
 )
 
-func typeOf[T any]() reflect.Type {
-	return reflect.TypeOf((*T)(nil)).Elem()
+type compiledQuery struct {
+	mask     ecs.Mask
+	exclude  ecs.Mask
+	Ids      []ecs.ID
+	compiled bool
 }
 
-func toIds(w *ecs.World, types []reflect.Type) []ecs.ID {
-	ids := make([]ecs.ID, len(types))
-	for i, t := range types {
-		ids[i] = ecs.TypeID(w, t)
+func (q *compiledQuery) Compile(w *ecs.World, include, optional, exclude []reflect.Type) {
+	if q.compiled {
+		return
 	}
-	return ids
+	q.Ids = toIds(w, include)
+	q.mask = toMaskOptional(w, q.Ids, optional)
+	q.exclude = toMask(w, exclude)
+	q.compiled = true
 }
 
-func toMask(w *ecs.World, types []reflect.Type) ecs.Mask {
-	mask := ecs.BitMask(0)
-	for _, t := range types {
-		mask.Set(ecs.TypeID(w, t), true)
+func (q *compiledQuery) MaskPair() ecs.MaskPair {
+	return ecs.MaskPair{
+		Mask:    q.mask,
+		Exclude: q.exclude,
 	}
-	return ecs.Mask{BitMask: mask}
-}
-
-func toMaskOptional(w *ecs.World, include []ecs.ID, optional []reflect.Type) ecs.Mask {
-	mask := ecs.NewBitMask(include...)
-	for _, t := range optional {
-		mask.Set(ecs.TypeID(w, t), false)
-	}
-	return ecs.Mask{BitMask: mask}
 }
 
 // Q0Builder builds a [Q0] query
 type Q0Builder struct {
-	include []reflect.Type
-	exclude []reflect.Type
+	include  []reflect.Type
+	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query0 creates a generic query for no components.
 //
 // See also [ecs.World.Query].
-func Query0() Q0Builder {
-	return Q0Builder{}
+func Query0() *Q0Builder {
+	return &Q0Builder{}
 }
 
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q0Builder) With(mask []reflect.Type) Q0Builder {
+func (q *Q0Builder) With(mask []reflect.Type) *Q0Builder {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -58,19 +55,16 @@ func (q Q0Builder) With(mask []reflect.Type) Q0Builder {
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q0Builder) Without(mask []reflect.Type) Q0Builder {
+func (q *Q0Builder) Without(mask []reflect.Type) *Q0Builder {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q0 query for iteration.
-func (q Q0Builder) Build(w *ecs.World) Q0 {
-	ids := toIds(w, q.include)
+func (q *Q0Builder) Build(w *ecs.World) Q0 {
+	q.compiled.Compile(w, q.include, []reflect.Type{}, q.exclude)
 	return Q0{
-		w.Query(ecs.MaskPair{
-			Mask:    ecs.Mask{BitMask: ecs.NewBitMask(ids...)},
-			Exclude: toMask(w, q.exclude),
-		}),
+		w.Query(q.compiled.MaskPair()),
 	}
 }
 
@@ -86,13 +80,14 @@ type Q1Builder[A any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query1 creates a generic query for one component.
 //
 // See also [ecs.World.Query].
-func Query1[A any]() Q1Builder[A] {
-	return Q1Builder[A]{
+func Query1[A any]() *Q1Builder[A] {
+	return &Q1Builder[A]{
 		include: []reflect.Type{typeOf[A]()},
 	}
 }
@@ -102,7 +97,7 @@ func Query1[A any]() Q1Builder[A] {
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q1Builder[A]) Optional(mask []reflect.Type) Q1Builder[A] {
+func (q *Q1Builder[A]) Optional(mask []reflect.Type) *Q1Builder[A] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -110,7 +105,7 @@ func (q Q1Builder[A]) Optional(mask []reflect.Type) Q1Builder[A] {
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q1Builder[A]) With(mask []reflect.Type) Q1Builder[A] {
+func (q *Q1Builder[A]) With(mask []reflect.Type) *Q1Builder[A] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -118,20 +113,17 @@ func (q Q1Builder[A]) With(mask []reflect.Type) Q1Builder[A] {
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q1Builder[A]) Without(mask []reflect.Type) Q1Builder[A] {
+func (q *Q1Builder[A]) Without(mask []reflect.Type) *Q1Builder[A] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q1 query for iteration.
-func (q Q1Builder[A]) Build(w *ecs.World) Q1[A] {
-	ids := toIds(w, q.include)
+func (q *Q1Builder[A]) Build(w *ecs.World) Q1[A] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q1[A]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids[0],
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids[0],
 	}
 }
 
@@ -155,13 +147,14 @@ type Q2Builder[A any, B any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query2 creates a generic query for two components.
 //
 // See also [ecs.World.Query].
-func Query2[A any, B any]() Q2Builder[A, B] {
-	return Q2Builder[A, B]{
+func Query2[A any, B any]() *Q2Builder[A, B] {
+	return &Q2Builder[A, B]{
 		include: []reflect.Type{typeOf[A](), typeOf[B]()},
 	}
 }
@@ -171,7 +164,7 @@ func Query2[A any, B any]() Q2Builder[A, B] {
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q2Builder[A, B]) Optional(mask []reflect.Type) Q2Builder[A, B] {
+func (q *Q2Builder[A, B]) Optional(mask []reflect.Type) *Q2Builder[A, B] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -179,7 +172,7 @@ func (q Q2Builder[A, B]) Optional(mask []reflect.Type) Q2Builder[A, B] {
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q2Builder[A, B]) With(mask []reflect.Type) Q2Builder[A, B] {
+func (q *Q2Builder[A, B]) With(mask []reflect.Type) *Q2Builder[A, B] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -187,20 +180,17 @@ func (q Q2Builder[A, B]) With(mask []reflect.Type) Q2Builder[A, B] {
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q2Builder[A, B]) Without(mask []reflect.Type) Q2Builder[A, B] {
+func (q *Q2Builder[A, B]) Without(mask []reflect.Type) *Q2Builder[A, B] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q2 query for iteration.
-func (q Q2Builder[A, B]) Build(w *ecs.World) Q2[A, B] {
-	ids := toIds(w, q.include)
+func (q *Q2Builder[A, B]) Build(w *ecs.World) Q2[A, B] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q2[A, B]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -234,13 +224,14 @@ type Q3Builder[A any, B any, C any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query3 creates a generic query for three components.
 //
 // See also [ecs.World.Query].
-func Query3[A any, B any, C any]() Q3Builder[A, B, C] {
-	return Q3Builder[A, B, C]{
+func Query3[A any, B any, C any]() *Q3Builder[A, B, C] {
+	return &Q3Builder[A, B, C]{
 		include: []reflect.Type{typeOf[A](), typeOf[B](), typeOf[C]()},
 	}
 }
@@ -250,7 +241,7 @@ func Query3[A any, B any, C any]() Q3Builder[A, B, C] {
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q3Builder[A, B, C]) Optional(mask []reflect.Type) Q3Builder[A, B, C] {
+func (q *Q3Builder[A, B, C]) Optional(mask []reflect.Type) *Q3Builder[A, B, C] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -258,7 +249,7 @@ func (q Q3Builder[A, B, C]) Optional(mask []reflect.Type) Q3Builder[A, B, C] {
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q3Builder[A, B, C]) With(mask []reflect.Type) Q3Builder[A, B, C] {
+func (q *Q3Builder[A, B, C]) With(mask []reflect.Type) *Q3Builder[A, B, C] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -266,20 +257,17 @@ func (q Q3Builder[A, B, C]) With(mask []reflect.Type) Q3Builder[A, B, C] {
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q3Builder[A, B, C]) Without(mask []reflect.Type) Q3Builder[A, B, C] {
+func (q *Q3Builder[A, B, C]) Without(mask []reflect.Type) *Q3Builder[A, B, C] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q3 query for iteration.
-func (q Q3Builder[A, B, C]) Build(w *ecs.World) Q3[A, B, C] {
-	ids := toIds(w, q.include)
+func (q *Q3Builder[A, B, C]) Build(w *ecs.World) Q3[A, B, C] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q3[A, B, C]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -318,13 +306,14 @@ type Q4Builder[A any, B any, C any, D any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query4 creates a generic query for four components.
 //
 // See also [ecs.World.Query].
-func Query4[A any, B any, C any, D any]() Q4Builder[A, B, C, D] {
-	return Q4Builder[A, B, C, D]{
+func Query4[A any, B any, C any, D any]() *Q4Builder[A, B, C, D] {
+	return &Q4Builder[A, B, C, D]{
 		include: []reflect.Type{typeOf[A](), typeOf[B](), typeOf[C](), typeOf[D]()},
 	}
 }
@@ -334,7 +323,7 @@ func Query4[A any, B any, C any, D any]() Q4Builder[A, B, C, D] {
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q4Builder[A, B, C, D]) Optional(mask []reflect.Type) Q4Builder[A, B, C, D] {
+func (q *Q4Builder[A, B, C, D]) Optional(mask []reflect.Type) *Q4Builder[A, B, C, D] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -342,7 +331,7 @@ func (q Q4Builder[A, B, C, D]) Optional(mask []reflect.Type) Q4Builder[A, B, C, 
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q4Builder[A, B, C, D]) With(mask []reflect.Type) Q4Builder[A, B, C, D] {
+func (q *Q4Builder[A, B, C, D]) With(mask []reflect.Type) *Q4Builder[A, B, C, D] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -350,20 +339,17 @@ func (q Q4Builder[A, B, C, D]) With(mask []reflect.Type) Q4Builder[A, B, C, D] {
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q4Builder[A, B, C, D]) Without(mask []reflect.Type) Q4Builder[A, B, C, D] {
+func (q *Q4Builder[A, B, C, D]) Without(mask []reflect.Type) *Q4Builder[A, B, C, D] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q4 query for iteration.
-func (q Q4Builder[A, B, C, D]) Build(w *ecs.World) Q4[A, B, C, D] {
-	ids := toIds(w, q.include)
+func (q *Q4Builder[A, B, C, D]) Build(w *ecs.World) Q4[A, B, C, D] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q4[A, B, C, D]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -410,13 +396,14 @@ type Q5Builder[A any, B any, C any, D any, E any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query5 creates a generic query for five components.
 //
 // See also [ecs.World.Query].
-func Query5[A any, B any, C any, D any, E any]() Q5Builder[A, B, C, D, E] {
-	return Q5Builder[A, B, C, D, E]{
+func Query5[A any, B any, C any, D any, E any]() *Q5Builder[A, B, C, D, E] {
+	return &Q5Builder[A, B, C, D, E]{
 		include: []reflect.Type{
 			typeOf[A](), typeOf[B](), typeOf[C](), typeOf[D](),
 			typeOf[E](),
@@ -429,7 +416,7 @@ func Query5[A any, B any, C any, D any, E any]() Q5Builder[A, B, C, D, E] {
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q5Builder[A, B, C, D, E]) Optional(mask []reflect.Type) Q5Builder[A, B, C, D, E] {
+func (q *Q5Builder[A, B, C, D, E]) Optional(mask []reflect.Type) *Q5Builder[A, B, C, D, E] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -437,7 +424,7 @@ func (q Q5Builder[A, B, C, D, E]) Optional(mask []reflect.Type) Q5Builder[A, B, 
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q5Builder[A, B, C, D, E]) With(mask []reflect.Type) Q5Builder[A, B, C, D, E] {
+func (q *Q5Builder[A, B, C, D, E]) With(mask []reflect.Type) *Q5Builder[A, B, C, D, E] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -445,20 +432,17 @@ func (q Q5Builder[A, B, C, D, E]) With(mask []reflect.Type) Q5Builder[A, B, C, D
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q5Builder[A, B, C, D, E]) Without(mask []reflect.Type) Q5Builder[A, B, C, D, E] {
+func (q *Q5Builder[A, B, C, D, E]) Without(mask []reflect.Type) *Q5Builder[A, B, C, D, E] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q5 query for iteration.
-func (q Q5Builder[A, B, C, D, E]) Build(w *ecs.World) Q5[A, B, C, D, E] {
-	ids := toIds(w, q.include)
+func (q *Q5Builder[A, B, C, D, E]) Build(w *ecs.World) Q5[A, B, C, D, E] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q5[A, B, C, D, E]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -511,13 +495,14 @@ type Q6Builder[A any, B any, C any, D any, E any, F any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query6 creates a generic query for six components.
 //
 // See also [ecs.World.Query].
-func Query6[A any, B any, C any, D any, E any, F any]() Q6Builder[A, B, C, D, E, F] {
-	return Q6Builder[A, B, C, D, E, F]{
+func Query6[A any, B any, C any, D any, E any, F any]() *Q6Builder[A, B, C, D, E, F] {
+	return &Q6Builder[A, B, C, D, E, F]{
 		include: []reflect.Type{
 			typeOf[A](), typeOf[B](), typeOf[C](), typeOf[D](),
 			typeOf[E](), typeOf[F](),
@@ -530,7 +515,7 @@ func Query6[A any, B any, C any, D any, E any, F any]() Q6Builder[A, B, C, D, E,
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q6Builder[A, B, C, D, E, F]) Optional(mask []reflect.Type) Q6Builder[A, B, C, D, E, F] {
+func (q *Q6Builder[A, B, C, D, E, F]) Optional(mask []reflect.Type) *Q6Builder[A, B, C, D, E, F] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -538,7 +523,7 @@ func (q Q6Builder[A, B, C, D, E, F]) Optional(mask []reflect.Type) Q6Builder[A, 
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q6Builder[A, B, C, D, E, F]) With(mask []reflect.Type) Q6Builder[A, B, C, D, E, F] {
+func (q *Q6Builder[A, B, C, D, E, F]) With(mask []reflect.Type) *Q6Builder[A, B, C, D, E, F] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -546,20 +531,17 @@ func (q Q6Builder[A, B, C, D, E, F]) With(mask []reflect.Type) Q6Builder[A, B, C
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q6Builder[A, B, C, D, E, F]) Without(mask []reflect.Type) Q6Builder[A, B, C, D, E, F] {
+func (q *Q6Builder[A, B, C, D, E, F]) Without(mask []reflect.Type) *Q6Builder[A, B, C, D, E, F] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q6 query for iteration.
-func (q Q6Builder[A, B, C, D, E, F]) Build(w *ecs.World) Q6[A, B, C, D, E, F] {
-	ids := toIds(w, q.include)
+func (q *Q6Builder[A, B, C, D, E, F]) Build(w *ecs.World) Q6[A, B, C, D, E, F] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q6[A, B, C, D, E, F]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -618,13 +600,14 @@ type Q7Builder[A any, B any, C any, D any, E any, F any, G any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query7 creates a generic query for seven components.
 //
 // See also [ecs.World.Query].
-func Query7[A any, B any, C any, D any, E any, F any, G any]() Q7Builder[A, B, C, D, E, F, G] {
-	return Q7Builder[A, B, C, D, E, F, G]{
+func Query7[A any, B any, C any, D any, E any, F any, G any]() *Q7Builder[A, B, C, D, E, F, G] {
+	return &Q7Builder[A, B, C, D, E, F, G]{
 		include: []reflect.Type{
 			typeOf[A](), typeOf[B](), typeOf[C](), typeOf[D](),
 			typeOf[E](), typeOf[F](), typeOf[G](),
@@ -637,7 +620,7 @@ func Query7[A any, B any, C any, D any, E any, F any, G any]() Q7Builder[A, B, C
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q7Builder[A, B, C, D, E, F, G]) Optional(mask []reflect.Type) Q7Builder[A, B, C, D, E, F, G] {
+func (q *Q7Builder[A, B, C, D, E, F, G]) Optional(mask []reflect.Type) *Q7Builder[A, B, C, D, E, F, G] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -645,7 +628,7 @@ func (q Q7Builder[A, B, C, D, E, F, G]) Optional(mask []reflect.Type) Q7Builder[
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q7Builder[A, B, C, D, E, F, G]) With(mask []reflect.Type) Q7Builder[A, B, C, D, E, F, G] {
+func (q *Q7Builder[A, B, C, D, E, F, G]) With(mask []reflect.Type) *Q7Builder[A, B, C, D, E, F, G] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -653,20 +636,17 @@ func (q Q7Builder[A, B, C, D, E, F, G]) With(mask []reflect.Type) Q7Builder[A, B
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q7Builder[A, B, C, D, E, F, G]) Without(mask []reflect.Type) Q7Builder[A, B, C, D, E, F, G] {
+func (q *Q7Builder[A, B, C, D, E, F, G]) Without(mask []reflect.Type) *Q7Builder[A, B, C, D, E, F, G] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q7 query for iteration.
-func (q Q7Builder[A, B, C, D, E, F, G]) Build(w *ecs.World) Q7[A, B, C, D, E, F, G] {
-	ids := toIds(w, q.include)
+func (q *Q7Builder[A, B, C, D, E, F, G]) Build(w *ecs.World) Q7[A, B, C, D, E, F, G] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q7[A, B, C, D, E, F, G]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }
 
@@ -731,13 +711,14 @@ type Q8Builder[A any, B any, C any, D any, E any, F any, G any, H any] struct {
 	include  []reflect.Type
 	optional []reflect.Type
 	exclude  []reflect.Type
+	compiled compiledQuery
 }
 
 // Query8 creates a generic query for eight components.
 //
 // See also [ecs.World.Query].
-func Query8[A any, B any, C any, D any, E any, F any, G any, H any]() Q8Builder[A, B, C, D, E, F, G, H] {
-	return Q8Builder[A, B, C, D, E, F, G, H]{
+func Query8[A any, B any, C any, D any, E any, F any, G any, H any]() *Q8Builder[A, B, C, D, E, F, G, H] {
+	return &Q8Builder[A, B, C, D, E, F, G, H]{
 		include: []reflect.Type{
 			typeOf[A](), typeOf[B](), typeOf[C](), typeOf[D](),
 			typeOf[E](), typeOf[F](), typeOf[G](), typeOf[H](),
@@ -750,7 +731,7 @@ func Query8[A any, B any, C any, D any, E any, F any, G any, H any]() Q8Builder[
 // Create the required mask with [Mask1], [Mask2], etc.
 //
 // Only affects component types that were specified in the query.
-func (q Q8Builder[A, B, C, D, E, F, G, H]) Optional(mask []reflect.Type) Q8Builder[A, B, C, D, E, F, G, H] {
+func (q *Q8Builder[A, B, C, D, E, F, G, H]) Optional(mask []reflect.Type) *Q8Builder[A, B, C, D, E, F, G, H] {
 	q.optional = append(q.optional, mask...)
 	return q
 }
@@ -758,7 +739,7 @@ func (q Q8Builder[A, B, C, D, E, F, G, H]) Optional(mask []reflect.Type) Q8Build
 // With adds more required components that are not accessible using Get... methods.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q8Builder[A, B, C, D, E, F, G, H]) With(mask []reflect.Type) Q8Builder[A, B, C, D, E, F, G, H] {
+func (q *Q8Builder[A, B, C, D, E, F, G, H]) With(mask []reflect.Type) *Q8Builder[A, B, C, D, E, F, G, H] {
 	q.include = append(q.include, mask...)
 	return q
 }
@@ -766,20 +747,17 @@ func (q Q8Builder[A, B, C, D, E, F, G, H]) With(mask []reflect.Type) Q8Builder[A
 // Without excludes entities with any of the given components from the query.
 //
 // Create the required mask with [Mask1], [Mask2], etc.
-func (q Q8Builder[A, B, C, D, E, F, G, H]) Without(mask []reflect.Type) Q8Builder[A, B, C, D, E, F, G, H] {
+func (q *Q8Builder[A, B, C, D, E, F, G, H]) Without(mask []reflect.Type) *Q8Builder[A, B, C, D, E, F, G, H] {
 	q.exclude = append(q.exclude, mask...)
 	return q
 }
 
 // Build builds a Q8 query for iteration.
-func (q Q8Builder[A, B, C, D, E, F, G, H]) Build(w *ecs.World) Q8[A, B, C, D, E, F, G, H] {
-	ids := toIds(w, q.include)
+func (q *Q8Builder[A, B, C, D, E, F, G, H]) Build(w *ecs.World) Q8[A, B, C, D, E, F, G, H] {
+	q.compiled.Compile(w, q.include, q.optional, q.exclude)
 	return Q8[A, B, C, D, E, F, G, H]{
-		w.Query(ecs.MaskPair{
-			Mask:    toMaskOptional(w, ids, q.optional),
-			Exclude: toMask(w, q.exclude),
-		}),
-		ids,
+		w.Query(q.compiled.MaskPair()),
+		q.compiled.Ids,
 	}
 }